+++ src/main/java/net/zaharenko424/casualties_cubed/client/gui/minigames/InjectMingameScreen.java
@@ -149,9 +149,9 @@ public class InjectMingameScreen extends Screen {
         Player player = Minecraft.getInstance().player;
         if (player != null) {
             Optional<Float> cons = player.getCapability(PlayerHealthProvider.PLAYER_HEALTH_DATA).map(PlayerHealthData::getConsciousness);
-            Optional<Double> pain = player.getCapability(PlayerHealthProvider.PLAYER_HEALTH_DATA).map(PlayerHealthData::getAveragePain);
+            Optional<Float> pain = player.getCapability(PlayerHealthProvider.PLAYER_HEALTH_DATA).map(PlayerHealthData::getAveragePain);
             float consscale = (cons.orElse(100f) / 100) * 0.15f;
-            float painscale = (float) (pain.orElse(0d) / 100);
+            float painscale = pain.orElse(0f) / 100;
             handObject.setShakeScale(painscale);
             handObject.setStiffness(consscale);
         }
